+ back Soutenance config

+++ src/main/java/iut/sae/saeconnectback/dtos/SoutenanceDTO.java
@@ -4,11 +4,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 import java.util.Date;
+import java.util.Optional;
 
 @Data
 public class SoutenanceDTO {
 
-    private int id;
+        private Long id;
 
     @JsonProperty(value = "sae_id")
     private Long idSae;
